% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasisFuncApproximation.R
\name{BasisFuncApproximation}
\alias{BasisFuncApproximation}
\title{Basis Function Approximation}
\usage{
BasisFuncApproximation(designMtrx, grid, degrFreedom, degree, intercept, knots)
}
\arguments{
\item{designMtrx}{the design matrix of time-invariant covariates.}

\item{grid}{a vector of the grid points for the basis function approximation (length equal to the number of rows of designMtrx).}

\item{degrFreedom}{the degrees of freedom for the polynomial basis used. See \code{\link{bs}} documentation.}

\item{degree}{the degree of the polynomial basis used, default is 3 which corresponds to cubic splines.}

\item{intercept}{either TRUE or FALSE. See bs documentation.}

\item{knots}{the position of interior knots used for the basis expansion, if NULL then the median is used. See bs documentation.}
}
\value{
The basis approximation matrix of the imported design matrix.
}
\description{
Takes a design matrix and converts it to the corresponding basis function approximation using B-splines.
}
