% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffFunctionEstimation.R
\name{coeffFunctionEstimation}
\alias{coeffFunctionEstimation}
\title{Coefficient function estimation}
\usage{
coeffFunctionEstimation(
  p,
  LgridA,
  gridA,
  degrFreedom,
  degree,
  intercept,
  b,
  knots
)
}
\arguments{
\item{p}{number of parameters in the data.}

\item{LgridA}{number of grid points on which the coefficient functions will be estimated.}

\item{gridA}{the grid points on which the coefficient functions will be estimated.}

\item{degrFreedom}{the degrees of freedom for the polynomial basis used. See \code{\link{bs}} documentation.}

\item{degree}{the degree of the polynomial basis used, default is 3 which corresponds to cubic splines. See \code{\link{bs}} documentation.}

\item{intercept}{either TRUE or FALSE. See \code{\link{bs}} documentation.}

\item{b}{the estimated coefficients' vector.}

\item{knots}{the position of interior knots used for the basis expansion, if NULL then the median is used. See \code{\link{bs}} documentation.}
}
\value{
The estimated coefficient functions.
}
\description{
Takes the estimated coefficients of the basis expansion and returns the estimated coefficient function for each covariate.
}
