% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbioSelect.R
\name{fbioSelect}
\alias{fbioSelect}
\title{Two-stage variable selection algorithm}
\usage{
fbioSelect(
  y,
  id,
  t.var,
  Z,
  X,
  low.dim.cov.number = 1,
  threshold.size,
  type.of.penalty,
  degree = 3,
  intercept = FALSE,
  degrFreedom = NULL,
  knots = NULL,
  weights = NULL,
  nfolds = 10,
  seed = 1000,
  ncores = 3
)
}
\arguments{
\item{y}{a response vector where only continuous responses are acceptable. Responses are expected to be time or dose varying.}

\item{id}{a vector including the identification codes for the response vector supplied.}

\item{t.var}{the varying time or dose under which the responses were recorded.}

\item{Z}{a matrix of low-dimensional covariates (should be an object obtained from the \code{\link{model.matrix}} function).}

\item{X}{a dataframe or matrix of high-dimensional covariates subject to variable selection.}

\item{low.dim.cov.number}{number of low-dimensional covariates, minimum is 1.}

\item{threshold.size}{threshold used for the variable screening step of the algorithm.}

\item{type.of.penalty}{type of penalty used for the second step of the algorithm (group penalised regression). Options of penalties can be found in the documentation of the \code{\link{grpreg}} documentation.}

\item{degree}{the degree of the polynomial basis used, default is 3 which corresponds to cubic splines.}

\item{intercept}{either TRUE or FALSE. See bs documentation. See \code{\link{bs}} documentation.}

\item{degrFreedom}{the degrees of freedom for the polynomial basis used. See \code{\link{bs}} documentation.}

\item{knots}{the position of interior knots used for the basis expansion, if NULL then the median is used. See \code{\link{bs}} documentation.}

\item{weights}{regression weights, if NULL then independence covariance structure is used for the repeated measures data.}

\item{nfolds}{number of folds for the cross-validation selecting the penalty size for the second step of the algorithm. See \code{\link{grpreg}} documentation.}

\item{seed}{the seed of R's random number generator. Default is 1000.}

\item{ncores}{number of cores used for parallel processing. Default is 3.}
}
\value{
A list including the ranking of the high dimensional covariates based on the marginal fit (variable screening step), the high-dimensional covariates that have been selected from the second step (group penalised regression), the estimated coefficient functions, the regression weights used and the group penalised regression path (\code{\link{grpreg}} output).
}
\description{
From a high-dimensional set of time-invariant covariates, selects those which are associated to the functional (repeated measures) response given some how-dimensional covariates.
}
