% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niFind.R
\name{niFind}
\alias{niFind}
\title{Compute regression weights for repeated measures data}
\usage{
niFind(data, idLabel)
}
\arguments{
\item{data}{a dataframe including the response vector.}

\item{idLabel}{the name of the variable in the dataframe containing subjects ids.}
}
\value{
Independence structure repeated measures weights.
}
\description{
Takes the data and subject ids and creates independence weights.
}
