% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{screening}
\alias{screening}
\title{Variable screening}
\usage{
screening(
  i,
  BasisMtrx,
  X,
  y,
  Ggroup,
  luniq_id,
  uniq_id,
  lut.var,
  ut.var,
  p,
  Lgrid,
  t.varGrid,
  degrFreedom,
  intercept,
  degree,
  knots,
  t.var,
  weights
)
}
\arguments{
\item{i}{iteration index.}

\item{BasisMtrx}{the basis approximation matrix of the low-dimensional covariate design matrix.}

\item{X}{a dataframe or matrix of high-dimensional covariates subject to variable selection.}

\item{y}{a response vector where only continuous responses are acceptable. Responses are expected to be time or dose varying.}

\item{Ggroup}{the groups created from the basis expansion in the BasisMtrx matrix.}

\item{luniq_id}{length of unique ids in the data.}

\item{uniq_id}{unique ids in the data.}

\item{lut.var}{length of unique time or dose points.}

\item{ut.var}{the unique time or dose points.}

\item{p}{number of parameters in the data.}

\item{Lgrid}{number of grid points on which the coefficient functions are estimated.}

\item{t.varGrid}{time of dose grid point on which the coefficient functions are estimated.}

\item{degrFreedom}{the degrees of freedom for the polynomial basis used. See \code{\link{bs}} documentation.}

\item{intercept}{either TRUE or FALSE. See \code{\link{bs}} documentation.}

\item{degree}{the degree of the polynomial basis used, default is 3 which corresponds to cubic splines.See \code{\link{bs}} documentation.}

\item{knots}{the position of interior knots used for the basis expansion, if NULL then the median is used. See \code{\link{bs}} documentation.}

\item{t.var}{the varying time or dose under which the responses were recorded.}

\item{weights}{regression weights.}
}
\value{
Te weighted mean square error of model fit.
}
\description{
Marginally fits a model including one of the high-dimensional covariates and the low-dimensional covariate matrix.
}
